Revise the psuedocode for the controller
The previous psuedocode relied on accepting user input for the console
version of the game, which is not necessary (and I suspect harder to
implement than what the new psuedocode entails).

This revised code we expect allows playing the game via executing
functions from the console, adhering to the advisory provided by the
project specs. We also made adjustments to simplify the logic and
improve modularity.