Change ending display messages to a dialog
Reaosning: It might be helpful to separate the alert box from the
results considering a game's conclusion has multiple elements we want to
display: a result message, a scoreline, and a 'play again' button.

The getPlayerScores function was modified to return more comprehensive
data about the players for usage in the ending dialog.