Refactor functions involved in playing a turn
Moved the handling of what error message to display from the gameControl
module to gameDisplay - it is a more logical place to put it.

In the process, made the return for gameBoard.placeMark more informative
particularly in the case of a player clicking on an already occupied
cell. A value of 'false' is a more semantic way to handle this behavior
than throwing an error.