Add partial styling to page & dialogs
A new function 'putMark' was added to script.js to place image symbols
instead of simply 'x's and 'o's.

Note that a side effect of using SVG images to display the mark symbols
is that it makes event delegation more difficult, so I chose to simply
add event listeners to all (9) cells; left a comment about this.