Fix mark collision not detected correctly
Click events on the cells being passed to handleCellClick would have
Event.target incorrectly be the SVG element if clicking on an already
occupied cell, causing the index lookup (with data-i and data-j) to
always fail at locating the cell actually being clicked. This causes
cell (0,0) to be checked for occupancy rather than the cell being
clicked, if different.