Improve gameDisplay logic separation & readability
More emphasis is placed on gameDisplay.update() to update alert messages
by providing it an optional message parameter. This change aims to
decouple the internal logic of calling gameControl and the external
logic of updating the display.

What prevents this from being called a refactor is that additional calls
to update() is less performant and has redundancy: it is called even
when the gameboard array has not changed, such as the case of clicking
on an occupied cell. This can be optimized by adding a flag argument to
update() that determines whether the game array should be rescanned.

Several other changes were made for better readability & brevity,
including clearer variable names, reordering, and restructuring.